[BACKEND] Refactor Ability model and exporter

+++ com.wsunitstats.exporter/src/main/java/com/wsunitstats/exporter/model/json/gameplay/submodel/ZoneEventJsonModel.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class ZoneEventJsonModel {
     private List<Integer> abilities;
     private Integer envSearchDistance;
-    private Integer envTags;
+    private Long envTags;
     private Integer levels;
     private Integer size;
 }
